refactor(crawler): extract FetchRouter to fetch_router.rs (strangler fig A)
Move FetchRouter enum, build_fetch_router(), and the Downloader impl
out of engine.rs into a dedicated fetch_router.rs module (strangler
fig corte A, issue #438). Pure code movement — zero behavior change.

- engine.rs: -164 lines; drops now-unused downloader imports
- mod.rs: adds pub(crate) mod fetch_router; repoints re-exports
- crawl_task_ctx.rs: import repointed engine:: -> fetch_router::
