+++ crates/webfang_core/src/application/crawler/mod.rs
@@ -8,6 +8,7 @@ pub mod concurrency_level;
 pub mod crawl_task_ctx;
 pub mod discovery;
 pub mod engine;
+pub(crate) mod fetch_router;
 pub mod progress;
 pub mod sitemap_discovery;
 
@@ -17,7 +18,6 @@ pub use discovery::{
     crawl_with_sitemap, discover_urls_for_tui, extract_content, parse_sitemap,
     scrape_single_url_for_tui,
 };
-pub use engine::{
-    build_fetch_router, crawl_site, crawl_site_with_options, EngineOptions, FetchRouter,
-};
+pub use engine::{crawl_site, crawl_site_with_options, EngineOptions};
+pub use fetch_router::{build_fetch_router, FetchRouter};
 pub use progress::CrawlProgress;
